+++ doc/changelog.rst
@@ -4,6 +4,28 @@ Changelog
 Changes in Version 3.13.0
 -------------------------
 
+MongoClient.max_bson_size/max_message_size/max_write_batch_size are deprecated.
+
+Deprecated :attr:`pymongo.mongo_client.MongoClient.max_bson_size`,
+:attr:`pymongo.mongo_client.MongoClient.max_message_size`, and
+:attr:`pymongo.mongo_client.MongoClient.max_write_batch_size`. These helpers
+were incorrect when in ``loadBalanced=true mode`` and ambiguous in clusters
+with mixed versions. Use the `hello command`_ to get the authoritative
+value from the remote server instead. Code like this::
+
+    max_bson_size = client.max_bson_size
+    max_message_size = client.max_message_size
+    max_write_batch_size = client.max_write_batch_size
+
+can be changed to this::
+
+    doc = client.admin.command('hello')
+    max_bson_size = doc['maxBsonObjectSize']
+    max_message_size = doc['maxMessageSizeBytes']
+    max_write_batch_size = doc['maxWriteBatchSize']
+
+.. _hello command: https://docs.mongodb.com/manual/reference/command/hello/
+
 Issues Resolved
 ...............
 
@@ -23,6 +45,9 @@ Deprecations
 - Deprecated :meth:`~pymongo.collection.Collection.map_reduce` and
   :meth:`~pymongo.collection.Collection.inline_map_reduce`.
   Use :meth:`~pymongo.collection.Collection.aggregate` instead.
+- Deprecated :attr:`pymongo.mongo_client.MongoClient.max_bson_size`.
+- Deprecated :attr:`pymongo.mongo_client.MongoClient.max_message_size`.
+- Deprecated :attr:`pymongo.mongo_client.MongoClient.max_write_batch_size`.
 
 See the `PyMongo 3.13.0 release notes in JIRA`_ for the list of resolved issues
 in this release.